refactor(test): extract shared getError helper to matchers/utils.ts
- Move getError helper from individual test files into a shared
  test/core/test/matchers/utils.ts module
- Fix test name: 'toEqual' → 'toStrictEqual' to match actual suggestion
- All 120 matcher tests still pass

Addresses Copilot review feedback on #10055.
