+++ test/core/test/matchers/toEqual.test.ts
@@ -1,27 +1,5 @@
-import { stripVTControlCharacters } from 'node:util'
-import { processError } from '@vitest/utils/error'
 import { describe, expect, it } from 'vitest'
-
-/**
- * Captures the error thrown by the provided function and returns a snapshot-friendly
- * object with the error message, diff, actual and expected values.
- * Part of #9074 — standardize matcher failure messages.
- */
-function getError(f: () => unknown) {
-  try {
-    f()
-  }
-  catch (error) {
-    const e = processError(error, { expand: true })
-    return {
-      message: stripVTControlCharacters(e.message),
-      diff: e.diff ? stripVTControlCharacters(e.diff) : undefined,
-      expected: e.expected,
-      actual: e.actual,
-    }
-  }
-  return expect.unreachable()
-}
+import { getError } from './utils'
 
 describe('toEqual', () => {
   it('passes when values are deeply equal', () => {
